sizet

+++ includes/item.h
@@ -51,7 +51,7 @@ typedef struct s_item{
 
 item *load_item(const char *filename);
 item **load_items(const char *dir, size_t *nItem);
-void freeItemList(item **items, int length);
+void freeItemList(item **items, size_t length);
 bool checkCraftValidity(item **items, int length);
 void freeItem(item *item);
 item *copyItem(item *i);
